Add two new options to the document class
Two new options for the document class are available, `lightheader` and
`nocolors`. The first will remove the dark background on the header, the
latter will remove all colors from the document. The old `print` option
is deprecated but the same effect can be obtained by turning on the
former two options.